requirements.txt
altgraph==0.17.4
colorama==0.4.6
iniconfig==2.0.0
MouseInfo==0.1.3
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
pip-review==1.3.0
pluggy==1.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.3
pytweening==1.2.0
pywin32==308
pywin32-ctypes==0.2.3
setuptools==75.3.0
tk==0.1.0
